Re-wrote keyboard controller.
The original implementation was a modification of the QEMU keyboard
code, with my own functions to hook it into the KVM system and take
care of other things. Here, I've tried to re-implement those QEMU
functions in a minimal way. The result is an extremely simple, yet
functional, keyboard system. It has been tested under Linux only.